requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn==0.30.6
psutil==6.0.0
requests==2.32.3
